refactor: extract faces routes into build_faces_router
Move all faces UI logic from faces_review_server.build_app into
webapp/routes_faces.py as a reusable APIRouter factory. The HTML
template uses __API_BASE__ placeholders so routes can be mounted at
any prefix. build_app delegates to build_faces_router(db, api_base="")
preserving existing root-mount behaviour unchanged.
