+++ pyproject.toml
@@ -82,3 +82,4 @@ select = ["E", "F", "W", "I"]
 "src/pyimgtag/review_server.py" = ["E501"]  # HTML/CSS template has long lines
 "src/pyimgtag/faces_review_server.py" = ["E501"]
 "src/pyimgtag/webapp/routes_review.py" = ["E501"]  # HTML/CSS template has long lines
+"src/pyimgtag/webapp/routes_faces.py" = ["E501"]  # HTML/CSS template has long lines
